Classrooms completed

+++ classes/views.py
@@ -1,15 +1,16 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
-from django.contrib.auth import login, authenticate, logout
-from .models import Classroom, Student
-from .forms import ClassroomForm, SignupForm, SigninForm, StudentForm
-
+from .forms import Signup,Signin
+from django.contrib.auth.models import User
+from django.contrib.auth import logout,login,authenticate
+from .models import Classroom,Student
+from .forms import ClassroomForm,StudentForm
 
 
 def signup(request):
-    form = SignupForm()
+    form = Signup()
     if request.method == 'POST':
-        form = SignupForm(request.POST)
+        form = Signup(request.POST)
         if form.is_valid():
             user = form.save(commit=False)
 
@@ -23,10 +24,11 @@ def signup(request):
     }
     return render(request, 'signup.html', context)
 
+
 def signin(request):
-    form = SigninForm()
+    form = Signin()
     if request.method == 'POST':
-        form = SigninForm(request.POST)
+        form = Signin(request.POST)
         if form.is_valid():
 
             username = form.cleaned_data['username']
@@ -41,14 +43,15 @@ def signin(request):
     }
     return render(request, 'signin.html', context)
 
-def signout(request):
-    logout(request)
-    return redirect("signin")
 
+def signout(request):
+	logout(request)
+	return render(request,'signin.html')
 
 
 
 def classroom_list(request):
+	
 	classrooms = Classroom.objects.all()
 	context = {
 		"classrooms": classrooms,
@@ -57,8 +60,9 @@ def classroom_list(request):
 
 
 def classroom_detail(request, classroom_id):
+	
 	classroom = Classroom.objects.get(id=classroom_id)
-	student=Student.objects.filter(classroom=classroom)
+	student=Student.objects.all().filter(classroom=classroom)
 	context = {
 		"classroom": classroom,
 		"student":student,
@@ -67,14 +71,16 @@ def classroom_detail(request, classroom_id):
 
 
 def classroom_create(request):
-
+	if not request.user.is_authenticated:
+		messages.warning(request, "Please signin")
+		return redirect('signin')
 	form = ClassroomForm()
 	if request.method == "POST":
 		form = ClassroomForm(request.POST, request.FILES or None)
 		if form.is_valid():
-			class_obj = form.save(commit=False)
-			class_obj.teacher = request.user
-			form.save()
+			Class=form.save(commit=False)
+			Class.teacher=request.user
+			Class.save()
 			messages.success(request, "Successfully Created!")
 			return redirect('classroom-list')
 		print (form.errors)
@@ -104,8 +110,10 @@ def classroom_update(request, classroom_id):
 		return render(request, 'update_classroom.html', context)
 
 
-
 def classroom_delete(request, classroom_id):
+	classroom = Classroom.objects.get(id=classroom_id)
+	if not (request.user.is_authenticated and request.user==classroom.teacher):
+		return redirect('warn')
 	Classroom.objects.get(id=classroom_id).delete()
 	messages.success(request, "Successfully Deleted!")
 	return redirect('classroom-list')
@@ -113,22 +121,66 @@ def classroom_delete(request, classroom_id):
 
 
 
-def add_student(request, classroom_id):
-	classroom_obj = Classroom.objects.get(id=classroom_id)
+def student_create(request,classroom_id):
+	classroom=Classroom.objects.get(id=classroom_id)
+	if not (request.user.is_authenticated and classroom.teacher==request.user):
+		return redirect('warn')
+	
 	form = StudentForm()
 	if request.method == "POST":
 		form = StudentForm(request.POST, request.FILES or None)
 		if form.is_valid():
-			student_obj = form.save(commit=False)
-			student_obj.classroom = classroom_obj
-			student_obj.save()
+			stu=form.save(commit=False)
+			stu.classroom=classroom
+			stu.save()
 			messages.success(request, "Successfully Added!")
-			return redirect('classroom-list')
+			return redirect('classroom-detail',classroom.id)
 		print (form.errors)
 	context = {
 	"form": form,
-	"class":classroom_obj
-
+	'class':classroom,
 	}
+
 	return render(request, 'add_student.html', context)
 
+
+
+
+def student_update(request, classroom_id,student_id):
+	
+	classroom=Classroom.objects.get(id=classroom_id)
+	if not (request.user.is_authenticated and classroom.teacher==request.user):
+		return redirect('warn')
+	student = Student.objects.get(id=student_id)
+	form = StudentForm(instance=student)
+	if request.method == "POST":
+		form = StudentForm(request.POST, request.FILES or None, instance=student)
+		if form.is_valid():
+			form.save()
+			messages.success(request, "Successfully Edited!")
+			return redirect('classroom-detail',classroom_id)
+		print (form.errors)
+	context = {
+	"form": form,
+	"student": student,
+	"classroom":classroom,
+	}
+
+	return render(request, 'update_student.html', context)
+
+
+
+
+def student_delete(request, classroom_id,student_id):
+	classroom=Classroom.objects.get(id=classroom_id)
+	if not (request.user.is_authenticated and classroom.teacher==request.user):
+		return redirect('warn')
+	stu=Student.objects.get(id=student_id)
+	stu.delete()
+	messages.success(request, "Successfully Deleted!")
+	return redirect('classroom-detail',classroom_id)
+
+def warning(request,):
+	
+	messages.warning(request, "invaid access!")
+	return render(request,"warning.html")
